+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.21.0
+- Updated styled components.
+- Removed unused component `EventHandler`.
+- Fixed rollup dependencies.
+
 # v0.20.2
 - Removed `gh-pages` package.